pyproject.toml
[tool.poetry]
name = "poc-redis-fastapi-chemblntd"
version = "0.1.1"
description = ""
authors = ["KB <kmbhm1@gmail.com>"]
readme = "README.md"
packages = [{include = "poc_redis_fastapi_chemblntd"}]

[tool.poetry.dependencies]
python = "^3.10"
aioredis = "2.0.0"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
pandas = "^2.0.3"
redis = "^4.6.0"
redis-om = "^0.2.1"
gunicorn = "^21.2.0"
chardet = "^5.1.0"
xlrd = "^2.0.1"
openpyxl = "^3.1.2"
fastapi-pagination = "^0.12.6"
msgpack = "^1.0.5"
rdkit = "^2023.3.2"
requests = "^2.31.0"
python-redis-lock = "^4.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
black = "^23.7.0"
isort = "^5.12.0"
pydocstyle = "^6.3.0"
pre-commit = "^3.3.3"
commitizen = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
ignore = [
    "F403",  # 'from module import *' used; unable to detect undefined names
    "F405",  # 'module' may be undefined, or defined from star imports: module
]
exclude = [
    # No need to traverse our git directory
    ".git",

    # There's no value in checking cache directories
    "__pycache__"
]

[tool.black]
line-length = 120
